refactor: siderBar -> MainSiderBar
