+++ src/components/ClubList.module.scss
@@ -3,6 +3,7 @@
   width: 300px !important;
 
   .logo {
+    height: 100px !important;
   }
 
   .recruit {
