+++ src/components/Main.tsx
@@ -5,13 +5,12 @@ import styles from "./Main.module.scss";
 import { Container, Row, Col } from "react-bootstrap";
 
 import LoginButton from "./LoginButton.module";
-import MainSideBar from "./SideBar";
+import MainSideBar from "./MainSideBar";
 import FilterClub from "./FilterClub";
 import { Route, Switch } from "react-router-dom";
 import UserAsk from "./UserAsk";
 import NotFound from "../pages/NotFound";
 import ClubList from "./ClubList";
-import ClubService from "../services/ClubService";
 
 interface MainProps {
   wishs: WishResType[] | null;
